+++ src/components/AreaChart/areachart.stories.tsx
@@ -1,5 +1,8 @@
+import React from "react"
 import type { Meta, StoryObj } from "@storybook/react"
 
+import { Label } from "../Label/Label"
+import { Switch } from "../Switch/Switch"
 import { AreaChart } from "./AreaChart"
 
 const chartdata = [
@@ -294,6 +297,30 @@ export const WithOnValueChange: Story = {
   },
 }
 
+export const WithTooltipCallback: Story = {
+  render: () => {
+    const [callback, setCallBack] = React.useState(null)
+    const [checked, setChecked] = React.useState(true)
+    return (
+      <>
+        <div className="flex items-center gap-3">
+          <Label htmlFor="a">showTooltip</Label>
+          <Switch id="a" checked={checked} onCheckedChange={setChecked} />
+        </div>
+
+        <AreaChart
+          data={chartdata}
+          index="date"
+          categories={["SolarCells", "Glass"]}
+          tooltipCallback={(props) => setCallBack(props.payload)}
+          showTooltip={checked}
+        />
+        <pre>{JSON.stringify(callback, null, 2)}</pre>
+      </>
+    )
+  },
+}
+
 export const WithLargeTickGap: Story = {
   args: {
     tickGap: 300,
@@ -328,5 +355,17 @@ export const OneDataValue: Story = {
     index: "date",
     categories: ["SolarCells", "Glass"],
     onValueChange: (v) => console.log(v),
+  },
+}
+
+export const WithFillSolid: Story = {
+  args: {
+    fill: "solid",
+  },
+}
+
+export const WithFillNone: Story = {
+  args: {
+    fill: "none",
   },
 }