+++ src/components/AreaChart/changelog.md
@@ -1,7 +1,26 @@
 # Tremor Raw AreaChart Changelog
 
+## 0.2.2
+
+### Changes
+
+- Fix: Legend scroll distance
+
+## 0.2.1
+
+### Changes
+
+- Fix: Gradient category id
+
+## 0.2.0
+
+### Changes
+
+- Feat: Add tooltipCallback prop
+- Feat: Add fill prop
+
 ## 0.1.0
 
 ### Changes
 
-Feat: Add legendPosition prop
+- Feat: Add legendPosition prop