+++ src/components/BarChart/changelog.md
@@ -1,5 +1,17 @@
 # Tremor Raw BarChart Changelog
 
+## 0.1.1
+
+### Changes
+
+- Fix: Legend scroll distance
+
+## 0.1.0
+
+### Changes
+
+- Feat: Add tooltipCallback prop
+
 ## 0.0.1
 
 ### Changes