+++ src/components/LineChart/changelog.md
@@ -1,7 +1,19 @@
 # Tremor Raw LineChart Changelog
 
+## 0.2.1
+
+### Changes
+
+- Fix: Legend scroll distance
+
+## 0.2.0
+
+### Changes
+
+- Feat: Add tooltipCallback prop
+
 ## 0.1.0
 
 ### Changes
 
-Feat: Added legendPosition prop
+- Feat: Added legendPosition prop