+++ src/components/SparkChart/changelog.md
@@ -1,7 +1,13 @@
 # Tremor Raw SparkChart Changelog
 
+## 0.1.1
+
+### Changes
+
+- Fix: Gradient category id
+
 ## 0.1.0
 
 ### Changes
 
-Feat: Add legendPosition prop
+- Feat: Add legendPosition prop